requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
dateparser==1.1.4
discord.py==2.1.0
Django==4.1.3
frozenlist==1.3.3
google-auth==2.14.1
google-auth-oauthlib==0.7.1
gspread==5.7.1
idna==3.4
multidict==6.0.2
oauthlib==3.2.2
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tweepy==4.12.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
whitenoise==6.2.0
yarl==1.8.1
